+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * added new function `minCellTypeStructDist`
 * added possibility to reconstruct everything around an object (`complement = TRUE`)
+* relaxed assertion of having at least two cells of each cell type in reconstruction to having at least one of one cell type
 
 # sosta 1.0.1